Migrate database to allow mfa relation to be nulll (#760)
When working with the mfa migration we found that it made the database
incompatable with old versions of the codebase. In an attempt to
preserve compatability the new migration makes the relations between the
existing models and the removed MFA models nullable in the database.
This allows both old and new code to interact with the new database
schema.