Update loggers with json.dumps for splunk ingestion
Found that Splunk was not ingesting the JSON for
log entries. The DICT type logging structures
need to be converted to JSON format for the
"message" part of the jsonout handler.

This improves the logging in Splunk, since
the fields are parsed and can be used in searches
rather than needing a REGEX.

This is needed in other areas of our logging too.

BB2-132